+++ alpha_autopilot/src/AutoPilot.cpp
@@ -25,14 +25,14 @@ void AutoPilot::update(){
     setpoint = automode->get_setpoint(state);
 
     pid_roll.set_setpoint(setpoint.rot.x);
-    double roll_effort = pid_roll.update(state.rot.x);
+    float roll_effort = pid_roll.update(state.rot.x);
     std::cout<<"roll setpoint "<<setpoint.rot.x<<std::endl;
     std::cout<<"roll state " <<state.rot.x<<std::endl;
     std::cout<<"roll effort "<<roll_effort<<std::endl;
     pid_pitch.set_setpoint(setpoint.rot.y);
-    double pitch_effort = pid_pitch.update(state.rot.y);
-    //    double throttle = automode->get_throttle();
-    double throttle = rc_in[THROTTLE_CH];//for debug
+    float pitch_effort = pid_pitch.update(state.rot.y);
+    //    float throttle = automode->get_throttle();
+    float throttle = rc_in[THROTTLE_CH];//for debug
     rc_out = compute_auto_rc_out(roll_effort,pitch_effort,throttle);//use trim 
     rc_out[ELEVATOR_CH] = rc_in[ELEVATOR_CH];//for debug
     //turn on the LED on ch5,and ch2,ch3,ch4,ch5 is only available
@@ -51,7 +51,7 @@ void AutoPilot::update(){
 
   publishRC(rc_out);
 }
-std::vector<int> AutoPilot::compute_auto_rc_out(double roll_effort,double pitch_effort,double throttle){
+std::vector<int> AutoPilot::compute_auto_rc_out(float roll_effort,float pitch_effort,float throttle){
   std::vector<int> rc_out = trim;
    rc_out[AUTOPILOT_LED_CH] = 4096;
     rc_out[THROTTLE_CH] = throttle;
